+++ pkg/api/getambassador.io/v2/zz_generated.conversion.go
@@ -947,7 +947,9 @@ func Convert_v3alpha1_AuthService_To_v2_AuthService(in *v3alpha1.AuthService, ou
 }
 
 func autoConvert_v2_AuthServiceIncludeBody_To_v3alpha1_AuthServiceIncludeBody(in *AuthServiceIncludeBody, out *v3alpha1.AuthServiceIncludeBody, s conversion.Scope) error {
-	*out = v3alpha1.AuthServiceIncludeBody(*in)
+	out.MaxBytes = in.MaxBytes
+	out.AllowPartial = in.AllowPartial
+	out.BodyAsRawBytes = in.BodyAsRawBytes
 	return nil
 }
 
@@ -957,7 +959,9 @@ func Convert_v2_AuthServiceIncludeBody_To_v3alpha1_AuthServiceIncludeBody(in *Au
 }
 
 func autoConvert_v3alpha1_AuthServiceIncludeBody_To_v2_AuthServiceIncludeBody(in *v3alpha1.AuthServiceIncludeBody, out *AuthServiceIncludeBody, s conversion.Scope) error {
-	*out = AuthServiceIncludeBody(*in)
+	out.MaxBytes = in.MaxBytes
+	out.AllowPartial = in.AllowPartial
+	out.BodyAsRawBytes = in.BodyAsRawBytes
 	return nil
 }