+++ pkg/api/getambassador.io/v3alpha1/crd_authservice.go
@@ -30,6 +30,8 @@ type AuthServiceIncludeBody struct {
 
 	// +kubebuilder:validation:Required
 	AllowPartial bool `json:"allow_partial,omitempty"`
+
+	BodyAsRawBytes *bool `json:"body_raw_bytes,omitempty"`
 }
 
 // TODO(lukeshu): In v3alpha2, consider getting rid of this struct type in favor of just using an