+++ CHANGELOG.md
@@ -87,6 +87,13 @@ Please see the [Envoy documentation](https://www.envoyproxy.io/docs/envoy/latest
 
 ## RELEASE NOTES
 
+## [2.2.1] TBD
+[2.2.1]: https://github.com/emissary-ingress/emissary/compare/v2.2.0...v2.2.1
+
+### Emissary-ingress and Ambassador Edge Stack
+
+- Bugfix: The Ambassador Agent now correctly supports requests to abort a rollout.
+
 ## [2.2.0] February 10, 2022
 [2.2.0]: https://github.com/emissary-ingress/emissary/compare/v2.1.2...v2.2.0